[too-few-public-methods] Use the multiple files template
