+++ doc/data/messages/t/too-few-public-methods/good/dataclass_and_function.py
@@ -1,0 +1,11 @@
+import dataclasses
+
+
+@dataclasses.dataclass
+class Worm:
+    name: str
+    fruit_of_residence: Fruit
+
+
+def bore(worm: Worm):
+    print(f"{worm.name} is boring into {worm.fruit_of_residence}")
