+++ doc/data/messages/t/too-few-public-methods/good/function.py
@@ -1,0 +1,2 @@
+def bore(fruit: Fruit, worm_name: str):
+    print(f"{worm_name} is boring into {fruit}")
