+++ doc/data/messages/t/too-few-public-methods/good/larger_api.py
@@ -1,6 +1,3 @@
-import dataclasses
-
-
 class Worm:
     def __init__(self, name: str, fruit_of_residence: Fruit):
         self.name = name
@@ -11,23 +8,3 @@ class Worm:
 
     def wiggle(self):
         print(f"{self.name} wiggle around wormily.")
-
-
-# or
-
-
-@dataclasses.dataclass
-class Worm:
-    name: str
-    fruit_of_residence: Fruit
-
-
-def bore(worm: Worm):
-    print(f"{worm.name} is boring into {worm.fruit_of_residence}")
-
-
-# or
-
-
-def bore(fruit: Fruit, worm_name: str):
-    print(f"{worm_name} is boring into {fruit}")
